Updated for 2017.1 - string changed to int ID - many internal naming and organizational changes - serialziation bugfixes

+++ Assets/URSA/Utility/ProjectInfo.cs
@@ -16,7 +16,7 @@
             }
         }
 
-        public float Version = 0.0f;
+        public int Version = 0;
     }
 
 }
